split large tests into smaller suites
